adopt mlx 0.29.1 and related mlx-c (#273)
- https://github.com/ml-explore/mlx/compare/v0.25.1...v0.29.1

**NOTE**

This change contains some breaking API changes in the area of quantization.  Specifically:

- the `quantized` / `dequantized` methods now take a `mode` parameter (not breaking)
- the `biases` result from `quantized` is now optional, e.g. `(wq: MLXArray, scales: MLXArray, biases: MLXArray?)`

We are keeping the same semver here to match with python mlx.  Although the change is breaking, it will likely be limited to implementations of quantized layers, e.g. `QuantizedLinear`, or other code that uses quantization directly.  `mlx-swift-examples` will have a synchronized release to reflect this change.

If you need to make a similar change, consider the changes from `QuantizedLinear`:

The properties changed from this:

```swift
    public let scales: MLXArray
    public let biases: MLXArray
```

to:

```swift
    public let mode: QuantizationMode
    public let scales: MLXArray
    public let biases: MLXArray?
```

A `mode` with parameter with a default value was added where needed: `mode: QuantizationMode = .affine` and the mode parameter was used in calls to the quantization APIs:

```swift
        var x = quantizedMatmul(
            x,
            weight,
            scales: scales,
            biases: biases,
            transpose: true,
            groupSize: groupSize,
            bits: bits,
            mode: mode
        )
```

and the `Quantizable` protocol was updated to have a `mode` parameter (protocol methods can't have default values):

```swift
    /// Return the module as a quantized representation
    func toQuantized(groupSize: Int, bits: Int, mode: QuantizationMode) -> Module
```